feat: daterangepicker/datepicker formatDate converted to momentjs, datepicker chevron justified to end -> making custom width possible

+++ components/daterangepicker/DateRangePicker.stories.tsx
@@ -20,7 +20,7 @@ export const Default: Story = {
         return (
             <DateRangePicker text={"Select a range"}
                              size={"medium"}
-                             dayFormat={"short"}
+                             dayFormat={"long"}
                              closeButton={true}
             />
         );
